+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - CASMCMS-9350: IMS Add create/modify/delete tests for images, recipes and public-keys
+- CASMCMS-9377: IMS: cmsdev: Verify delete operations
 
 ## [1.28.0] - 2025-03-31